funciones GCP/read_datastore_data/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore
flask